[top_processor] Regexp fix #9
Fixes variable pid size issues.  Apparently pids can go over 65k, which
I thought was the standard limit, but apparently not.  To fix, this adds
an option, --pid-size, to set the digit size of the pids in top.  This
defaults to 5, which is what I am used to, and is probably the most
common use case

And you might me:

  "Wait... Nick, there is no commit for Regexp fixes 1-8?"

Oh, there were at least 8 regexp fixes (probably more, but I was too
lazy to get an exact coung), but they made it into the initial commit.