[top_processor] Address timezone offset issues
So this is two fold:

1. There was a change in ruby 2.2 that I forgot about that changed how
timezone offsets were treated with `Time.parse`.  In ruby-2.2, this
causes Time.parse to not covert the time to localtime, which is what it
did in prior versions of ruby.  A helper method has been added to
address this to replicate the previous implementation on newer versions
of ruby.
2. The offset handling code was previously hardcoded to work with
Pacific timezones only, and this has been updated to make it
configurable via the `--offset` cli flag.