[miq_server_leak_discovery] Add YAML and sync_needed?
Adds two different scripts for testing different parts of the monitor
loop in MiqServer

04_yaml_dump_loop_test:  Tests part of `$log.log_hashes` which dumps the
settings to a hash.  The C based part of YAML dump made it suspect, but
after further research into the codebase, this portion isn't reached on
a regular basis.

05_miq_server_sync_needed_loop_test:  Tests the `sync_needed?` portion
of the monitor loop, which is a part of `MiqServer#monitor_workers`.
Currently showing a small bit of promise in being at least a cause of
the leak, but even if it does leak, seems like it is only a part of it.