Adds 10_miq_process_loop_test.rb
This test runs a consistent loop that tests the MiqProcess.processInfo
method, and specifically the code as how it is used in the application
when calling `MiqWorker.status_update_all`.  This is meant to be run in
the same directory as the `manageiq`/`cfme` application.

Of note:  This seems to leak on a consistent basis, but only when the
first line is present.  If the `require "config/environment"` line is
commented out, this will not leak.

This leads me (and others) to believe that the memory has something to
do with some core extension overrides in one of the 200+ gems in the
application, and a method happens to be calling one of those patched
methods here, causing the leak.