[leak_discovery] Test `require 'linux_admin'`
When testing MiqSystem.num_cpus, we found that it was leaking pretty
quick, even though most of what it was doing was running a quick `if` to
check if it was one the proper platform, a deferred `require` for
`linux_admin` (should normally be a no-op), and assigning a `memoized`
variable.

Turns out the `require` statement was actually the guilty party, and was
causing the leak.  This script is a quick way of testing and confirming
that leak.