[leak_discovery] Toggle require config/environment
Since the leak seems to be related to our environment (in test scripts,
the requiring of "config/environment" to setup the Rails env), this
makes it so that it can easilly be compared with and without by passing
the `WITHOUT_MIQ_ENV` env var.

**WITH**

    $ ruby -I.  /path/to/miq_tools/miq_server_leak_discovery/17_test_require_linux_admin.rb

**WITHOUT**

    $ WITHOUT_MIQ_ENV=1 bundle exec ruby /path/to/miq_tools/miq_server_leak_discovery/17_test_require_linux_admin.rb

Obviously, this must be run from the `manageiq` dir for the first form.
Kept the pathing consistent in the examples though between both forms.