[util] Add sync_dumps_and_logs script
Used to sync the evm.log, top_output.log, and any dump files that might
be in the tmp/ dir of the appliance.

This also takes care of copying the existing evm.log and top_output.log
and created partial versions of those, gzip'ing them prior to transfer,
and gzip'ing any dump files that are currently sitting raw in the tmp
dir.  Also, if the '-i' option is passed, it will sort the files further
to a subdirectory based on the hostname.

It utilizes `expect` so the user doesn't have to enter a username and
password for any of the ssh or rsync commands (might make this
toggleable in the future for those who upload their ssh keys to the
server... I on the other hand am too lazy for that noise).