[memory_leak_worker_logging] BUG: pid option to_s
Passing in the type of `Integer` into OptionParser#on will actually
convert it to an int (even though this was already being done.  But,
when comparing the `option[:pid]` to the Regexp match from the line, it
would always miss since the `options[:pid]` was a Integer, and the match
value was a string.

Since it happens only at the beginning of the script once, and we
currently have no use for the PID being a number, convert it to a string
so we can compare it directly to the match value.

The Interger validation is still being used in the OptionParser#on
portion, though, since this adds some nicer validation to the option
parser.