[util] Add util/multi_file_log_parser
MultiFileLogParser is meant to be a generic way of parsing files with
multiple outputs, and potentially multiple line matches... so basically
I have implemented a form of `sed`...

Basically, this allows you to assign multiple REGEXP to match on, setup
with named captures, and define a standardized way of defining multiple
output files based on an `id` field (configurable) and a datestamp (less
configurable currently).

The matches of each REGEXP are kept around in a buffer so other matches
can also use them, and each REGEXP can pass a block to manipulate the
data further, or decide when to write to the configured file.