Add seed_scripts/notifications
Adds a script to build a collection of notifications for purge testing
at scale.

Takes about 30 seconds to run with the default counts, which creates
roughly half a million records, and 5 minutes if you increase the --num
value to 100k (which creates 5 million records in total for all of our
default notification types).

TODO:  Add NotificationRecipient records as well.