[bisect_gemfile_parser] Always use `#.hours.ago`
Previously, we were using `git` to generate the friendly timestamp for
us to use in a commit.  Specifically, it would generate something like
the examples below:

    2.minutes.ago
    5.hours.ago
    3.weeks.ago
    1.month.ago

But the farther back in time we go, the less precision we have, so we
end up with the same repo commits after a while.

This change updates the code to figure out the "time ago" in hours, so
that we are always using relatively usable precision, without being too
precise.

Note:  Curious if it makes sense just to use seconds, and avoid the
conversion to hours.  Might make the commits more stable, but unsure.