[bisect_gemfile_parser] Only use merge commits
Since the only relevant commits we care about in the case of this parser
is the merge commits, make sure we are using those determining the
commit to use in the `bundler.d/bisect.rb`.

This is because the only difference that will be observed by the
manageiq repo normally will be the merge commits, so referencing other
commits will only put the bisecting into a weird state.