[MultiFileLogParser] Allow single file output
For now, this just allows for passing in a `:single_file` for the
options, which overrides the default functionality for writing files
(which uses a `:id_col` to help determine file name).

The previous form is still valid as well, just allows for passing in a
IO object.  This doesn't open the file for you (however), but that might
change in the future.