Add request_log_parser
This is a parser of the rails logger, and allows you to determine the
length of a given request by matching up the thread ID with the
"Started" and "Completed in" lines.

Also does some aggregation as well, and makes use of the
MultiFileLogParser to handle the two log lines it is searching for.

Built from:

https://gist.github.com/NickLaMuro/18b2fae94b26444e49f03920799f083c