[WIP] Add rspec_profiling_tools
Collection of scripts used to profile both rspec startup and individual
specs to determine slowness hotspots in each.

Marking this as [WIP] since I am blanket committing this for now to push
this up somewhere so I don't loose it as I am about to start working on
something else unrelated.