[har_file_parser] UiConstants support
In older versions of MIQ, `UiConstants` was a global constant is
required to be autoloaded for many of the controller actions to work.

Unfortunately, this was not done by default for the `rails runner` in
the older versions of MIQ where this existed, so this adds an option to
load it via the `miq_har_file_parser/cli`.