[har_file_parser] truely optional HAR_FILE ARGV
The help docs for the CLI allow for either passing a file.har as a final
argument, or as a command line argument via the `--input` flag.

Do to the way it was coded, it always forced using the result of the
`HAR_FILE` ARGV instead of the parsed option.  This change allows the
arguments to work as expected.