[har_file_parser] support usage on an appliance
NOTE:  MANY HOURS WASTED FIGURING THIS OUT!!1!

Turns out that the rails session will automatically log you out if you
are not in the same mode "http mode" (`http` or `https`), and this is
not super obvious when using the integration test framework to run
commands that this is happening (basically, stuff just starts
redirecting to login on you, with no logging at all).

Since all appliances are configured out of the box to use `https` in
puma, this seemed like a safe default in the script to just enable
`.https!` if we are on an appliance by just querying it.  Might make it
configurable in the future, but no promises.