[har_file_parser] Sort by startedDateTime
Turns out that Chrome and Firefox sort the "entries" differently.  The
Firefox (sanely) is in chronological order, but the Chrome one (for
whatever reason...) is not.

Sorting by `startedDateTime` ensures that we are get the results in the
order we requested them, allowing the generated script to function as
expected.

The `startedDateTime` is also added to the parsed requests hash, and
displayed in the summary printout.