Fix --migrate/--drop
Using `FalseClass` to define the class type for `.on` caused the flag to
always be `false`, even when the positive case was being used.  The
intent for doing that is probably when you want to do something like
this:

    opt.on "--enable",  TrueClass,  "Set to true"
    opt.on "--disable", FalseClass, "Set to false"

Which will define what the flag is everytime, but doesn't allow it to
change forms with `--[no-]*`.