[vmdb_db_restore] Add import_db
Allows using "textual style" database dumps (gzip'd or raw) as well as
`pg_restore` valid dumps.

Makes use of `Open3.popen2e` to allow for a block to be passed that will
be the contents of the `input` thread, and passes it the `stdin` object
so it can pipe data to the command given.  In this case:

    psql -U root -d DBNAME

Which will take anything from `stdin` as input and start processing it.
The data is unzipped via `zlib` if necessary, otherwise the raw data
will work as well..